+++ Manager/DeveloperForm.Designer.cs
@@ -34,6 +34,7 @@ namespace TonePrism.Manager
             this.txtFirstName = new System.Windows.Forms.TextBox();
             this.lblGrade = new System.Windows.Forms.Label();
             this.numGrade = new System.Windows.Forms.NumericUpDown();
+            this.chkGradeTeacher = new System.Windows.Forms.CheckBox();
             this.chkGradeUnknown = new System.Windows.Forms.CheckBox();
             this.btnOK = new System.Windows.Forms.Button();
             this.btnCancel = new System.Windows.Forms.Button();
@@ -79,7 +80,7 @@ namespace TonePrism.Manager
             this.lblGrade.Name = "lblGrade";
             this.lblGrade.Size = new System.Drawing.Size(100, 15);
             this.lblGrade.TabIndex = 4;
-            this.lblGrade.Text = "期生（0で教員）";
+            this.lblGrade.Text = "期生";
             // 
             // numGrade
             // 
@@ -90,7 +91,7 @@ namespace TonePrism.Manager
             0,
             0});
             this.numGrade.Minimum = new decimal(new int[] {
-            0,
+            1,
             0,
             0,
             0});
@@ -103,13 +104,24 @@ namespace TonePrism.Manager
             0,
             0});
             //
+            // chkGradeTeacher
+            //
+            this.chkGradeTeacher.AutoSize = true;
+            this.chkGradeTeacher.Location = new System.Drawing.Point(270, 70);
+            this.chkGradeTeacher.Name = "chkGradeTeacher";
+            this.chkGradeTeacher.Size = new System.Drawing.Size(53, 19);
+            this.chkGradeTeacher.TabIndex = 6;
+            this.chkGradeTeacher.Text = "教員";
+            this.chkGradeTeacher.UseVisualStyleBackColor = true;
+            this.chkGradeTeacher.CheckedChanged += new System.EventHandler(this.chkGradeTeacher_CheckedChanged);
+            //
             // chkGradeUnknown
             //
             this.chkGradeUnknown.AutoSize = true;
-            this.chkGradeUnknown.Location = new System.Drawing.Point(270, 70);
+            this.chkGradeUnknown.Location = new System.Drawing.Point(345, 70);
             this.chkGradeUnknown.Name = "chkGradeUnknown";
             this.chkGradeUnknown.Size = new System.Drawing.Size(53, 19);
-            this.chkGradeUnknown.TabIndex = 6;
+            this.chkGradeUnknown.TabIndex = 7;
             this.chkGradeUnknown.Text = "不明";
             this.chkGradeUnknown.UseVisualStyleBackColor = true;
             this.chkGradeUnknown.CheckedChanged += new System.EventHandler(this.chkGradeUnknown_CheckedChanged);
@@ -146,6 +158,7 @@ namespace TonePrism.Manager
             this.ClientSize = new System.Drawing.Size(450, 152);
             this.Controls.Add(this.btnCancel);
             this.Controls.Add(this.btnOK);
+            this.Controls.Add(this.chkGradeTeacher);
             this.Controls.Add(this.chkGradeUnknown);
             this.Controls.Add(this.numGrade);
             this.Controls.Add(this.lblGrade);
@@ -174,6 +187,7 @@ namespace TonePrism.Manager
         private System.Windows.Forms.TextBox txtFirstName;
         private System.Windows.Forms.Label lblGrade;
         private System.Windows.Forms.NumericUpDown numGrade;
+        private System.Windows.Forms.CheckBox chkGradeTeacher;
         private System.Windows.Forms.CheckBox chkGradeUnknown;
         private System.Windows.Forms.Button btnOK;
         private System.Windows.Forms.Button btnCancel;